+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/query/IMetadataQueryDAO.java
@@ -79,7 +79,7 @@ public interface IMetadataQueryDAO extends DAO {
     /**
      * @param agentId the agent id of the process.
      */
-    List<Process> listProcesses(final String agentId) throws IOException;
+    List<Process> listProcesses(final String agentId, long startPingTimeBucket, long endPingTimeBucket) throws IOException;
 
     /**
      * @param serviceId the service id of the process