+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/common/dao/JDBCMetadataQueryDAO.java
@@ -256,16 +256,7 @@ public class JDBCMetadataQueryDAO implements IMetadataQueryDAO {
     @Override
     @SneakyThrows
     public List<Process> listProcesses(String serviceId, ProfilingSupportStatus supportStatus, long lastPingStartTimeBucket, long lastPingEndTimeBucket) {
-        List<String> tables;
-        if (lastPingStartTimeBucket > 0 && lastPingEndTimeBucket > 0) {
-            tables = tableHelper.getTablesForRead(
-                ProcessTraffic.INDEX_NAME,
-                lastPingStartTimeBucket,
-                lastPingEndTimeBucket
-            );
-        } else {
-            tables = tableHelper.getTablesWithinTTL(ProcessTraffic.INDEX_NAME);
-        }
+        List<String> tables = tableHelper.getTablesWithinTTL(ProcessTraffic.INDEX_NAME);
         final var results = new ArrayList<Process>();
 
         for (String table : tables) {
@@ -305,11 +296,7 @@ public class JDBCMetadataQueryDAO implements IMetadataQueryDAO {
     @Override
     @SneakyThrows
     public List<Process> listProcesses(String serviceInstanceId, Duration duration, boolean includeVirtual) {
-        final var tables = tableHelper.getTablesForRead(
-            ProcessTraffic.INDEX_NAME,
-            duration.getStartTimeBucket(),
-            duration.getEndTimeBucket()
-        );
+        final List<String> tables = tableHelper.getTablesWithinTTL(ProcessTraffic.INDEX_NAME);
         final var results = new ArrayList<Process>();
 
         for (String table : tables) {
@@ -345,7 +332,7 @@ public class JDBCMetadataQueryDAO implements IMetadataQueryDAO {
 
     @Override
     @SneakyThrows
-    public List<Process> listProcesses(String agentId) {
+    public List<Process> listProcesses(String agentId, long startPingTimeBucket, long endPingTimeBucket) {
         final var tables = tableHelper.getTablesWithinTTL(ProcessTraffic.INDEX_NAME);
         final var results = new ArrayList<Process>();
 
@@ -372,11 +359,7 @@ public class JDBCMetadataQueryDAO implements IMetadataQueryDAO {
     @Override
     @SneakyThrows
     public long getProcessCount(String serviceId, ProfilingSupportStatus profilingSupportStatus, long lastPingStartTimeBucket, long lastPingEndTimeBucket) {
-        final var tables = tableHelper.getTablesForRead(
-            ProcessTraffic.INDEX_NAME,
-            lastPingStartTimeBucket,
-            lastPingEndTimeBucket
-        );
+        final var tables = tableHelper.getTablesWithinTTL(ProcessTraffic.INDEX_NAME);
         long total = 0;
 
         for (String table : tables) {